requirements.txt
Flask==0.10.1
Flask-Cors==3.0.2
gunicorn==19.6.0
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
requests==2.11.1
itsdangerous==0.23

numpy==1.11.2
opencv-python==3.2.0.7
Pillow==3.3.1